rename movie-stakeholders components and create folders pages/components
