+++ libs/movie/src/lib/stakeholder/components/stakeholder-item/stakeholder-item.component.scss
