+++ libs/movie/src/lib/stakeholder/components/stakeholder-repertory/stakeholder-repertory.component.scss
