--- libs/movie/src/lib/stakeholder/item/item.component.ts
@@ -1,19 +1,0 @@
-import { Component, ChangeDetectionStrategy, Input } from '@angular/core';
-import { Stakeholder, StakeholderService } from '../+state';
-import { MovieQuery } from '../../movie/+state';
-
-@Component({
-  selector: 'stakeholder-item',
-  templateUrl: './item.component.html',
-  styleUrls: ['./item.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-export class StakeholderItemComponent {
-  @Input() stakeholder: Stakeholder;
-
-  constructor(private service: StakeholderService, private movieQuery: MovieQuery,) { }
-
-  public removeStakeholder() {
-    this.service.remove(this.movieQuery.getActiveId(), this.stakeholder.id);
-  }
-}
