+++ libs/movie/src/lib/stakeholder/pages/stakeholder-view/stakeholder-view.component.scss
